fix(hotkeys): preserve hotkeys if changed, and reduce re-rendering
